consolidate linting code
